requirements.txt
black==19.10b0
numpy==1.22.1
pandas==1.4.2
pip==21.2.4
pylint==2.12.2
pytest==7.1.1
torch==1.11.0
torchvision==0.12.0
monai-weekly==0.9.dev2222
monailabel-weekly==0.4.dev2222
psutil==5.8.0
GeodisTK
itk==5.2.1.post1
scikit-image==0.19.2
pydicom==2.2.2
matplotlib==3.5.2
opencv-python~=4.5.5
SimpleCRF
dash
